+++ docs/plotDistributionsExamples.R
@@ -1,42 +1,44 @@
 # plotDistributions
-## Distribution Plots in the Package
+## Density and Cumulative Distribution Examples
 
+### Normal Distribution
 
-### Evaluates lower tail with equal to inequality.
-plotBinom(stat = 1, size = 3, prob = 0.5, section = "lower", strict = FALSE) 
+norm.pdf()
+norm.pdf(limits=c(1,2))
+norm.pdf(probs=c(.025,.975))
+norm.pdf(probs=c(.5,.9999))
 
-### Evaluates bounded region with lower bound equal to and upper bound strict inequality.
-plotBinom(stat = c(1,2), size = 5, prob = 0.35, section = "bounded", strict = c(0,1))
+norm.cdf()
+norm.cdf(limits=c(1,2))
+norm.cdf(probs=c(.025,.975))
 
-### Evaluates upper tail with strict inequality.
-plotBinom(stat = 1, size = 3, prob = 0.5, section = "upper", strict = TRUE)
+### Student's Distribution
 
+t.pdf(params=c(df=5))
+t.pdf(params=c(df=5),limits=c(1,2))
+t.pdf(params=c(df=5),probs=c(.025,.975))
 
-### Evaluates lower tail.
-plotNorm(stat = 1, mu = 4, sd = 5, section = "lower") 
+t.cdf(params=c(df=5))
+t.cdf(params=c(df=5),limits=c(1,2))
+t.cdf(params=c(df=5),probs=c(.025,.975))
 
-### Evaluates bounded region.
-plotNorm(stat = c(3,7), mu = 5, sd = 3, section = "bounded")
+### F Distribution
 
-### Evaluates upper tail.
-plotNorm(stat = 1, mu = 3, sd = 2, section = "upper")
+f.pdf(params=c(df1=2,df2=10))
+f.pdf(params=c(df1=2,df2=10),limits=c(1,1000))
+f.pdf(params=c(df1=2,df2=10),probs=c(.95,.99999))
 
+f.cdf(params=c(df1=2,df2=10))
+f.cdf(params=c(df1=2,df2=10),limits=c(1,1000))
+f.cdf(params=c(df1=2,df2=10),probs=c(.95,.99999))
 
-### Evaluates lower tail.
-plotT(stat = -1, df = 3, section = "lower") 
+### Studentized Range Distribution
 
-### Evaluates bounded region.
-plotT(stat = c(-2.13,2.13), df = 15, section = "bounded") 
+q.pdf(params=c(nmeans=2,df=10))
+q.pdf(params=c(nmeans=2,df=10),limits=c(1,1000))
+q.pdf(params=c(nmeans=2,df=10),probs=c(.95,.99999))
 
-### Evaluates upper tail.
-plotT(stat = 1, df = 15, section = "upper") 
+q.cdf(params=c(nmeans=2,df=10))
+q.cdf(params=c(nmeans=2,df=10),limits=c(1,1000))
+q.cdf(params=c(nmeans=2,df=10),probs=c(.95,.99999))
 
-
-### Evaluates lower tail.
-plotF(stat = 1, df1 = 2, df2 = 5, section = "lower") 
-
-### Evaluates bounded region.
-plotF(stat = c(3,5), df1 = 2, df2 = 5, section = "bounded") 
-
-### Evaluates upper tail.
-plotF(stat = 3.4, df1 = 2, df2 = 5, section = "upper")