+++ NEWS.md
@@ -3,3 +3,6 @@
 ## Developmental Version
 
 - Initial release
+- Plotting of built in R distributions
+- Added probability density for the Studentized Range Distribution
+- Examples of distributions with options